requirements.txt
boto3==1.17.69
mysql_connector_repackaged==0.3.1
